RavenDB-22635 - Reading the snapshot from stream isn't under the tx write lock anymore
