+++ src/Raven.Server/ServerWide/Context/ClusterContextPool.cs
@@ -9,6 +9,7 @@ namespace Raven.Server.ServerWide.Context
     {
         private ClusterChanges _changes;
         private StorageEnvironment _storageEnvironment;
+        public StorageEnvironment StorageEnvironment => _storageEnvironment;
 
         public ClusterContextPool(ClusterChanges changes, StorageEnvironment storageEnvironment, Size? maxContextSizeToKeepInMb = null)
             : base(maxContextSizeToKeepInMb)
